Wsrast patch 1 (#1)
* Change devDependencies to peerDependencies

This will allow a call to install-peerdeps to prevent the need to copy/paste from the readme.md. Instead, a single command can be used.

* Update readme with new instructions

* Update with install-peerdeps call method

Add a way to use the `install-peerdeps` package to do all the
manual installs, rather than relying on copy/paste. Document
changes. Make publishable to NPM.

* Add package meta data

* Update npx command and remove second parameter

* Bump version 1.0.1