Refactor the way files are added
Getting rid of the _curlrc file and going with a
node-based solution in request.js that runs on the
post-install NPM event, so that the user doesn't
have to add the files by hand.
Moving the linting config files to the root folder
so that they can also be used during development
of this package.