Add ESLint jest plugin and rule mods
I'm adding the trailingCommas setting for prettier,
as there's good reason to have this there: easy
additions/removals and it cleans up git diff messages.

I'm also adding the ESLint plugin for Jest to try and
take care of the "describe", "it" and "expect" functions
being highlighted as errors in the parent project.